mesh_and_shape

+++ src/main/java/shapes/Shape.java
@@ -2,5 +2,5 @@ package shapes;
 
 public interface Shape {
     public float[] getVertices();
-    public float[] getIndices();
+    public int[] getIndices();
 }
